+++ packages/myst-to-html/package.json
@@ -43,8 +43,7 @@
     "mdast": "^3.0.0",
     "mdast-util-find-and-replace": "^2.1.0",
     "mdast-util-to-hast": "^12.3.0",
-    "myst-common": "^1.1.2",
-    "myst-transforms": "^1.1.0",
+    "myst-common": "^1.1.4",
     "rehype-format": "^4.0.1",
     "rehype-parse": "^8.0.4",
     "rehype-remark": "^9.1.2",
