--- packages/myst-transforms/src/types.ts
@@ -1,14 +1,0 @@
-// TODO: move this to myst-spec
-export enum AdmonitionKind {
-  admonition = 'admonition',
-  attention = 'attention',
-  caution = 'caution',
-  danger = 'danger',
-  error = 'error',
-  important = 'important',
-  hint = 'hint',
-  note = 'note',
-  seealso = 'seealso',
-  tip = 'tip',
-  warning = 'warning',
-}
